backend/requirements.txt
fastapi
uvicorn
python-multipart
faster-whisper
fugashi
unidic-lite
requests
ffmpeg-python
celery
redis
python-dotenv
pydantic
sqlalchemy
pydub
jinja2
python-slugify
httpx
transformers
torch
sentencepiece
jamdict
